[GREEN] Returns Jobs filtered by search radius after rebase main
