+++ src/integrationTest/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/controller/jobs/PostcodeMother.kt
@@ -11,6 +11,22 @@ object PostcodeMother {
 
   const val RELEASE_AREA_POSTCODE = "AG121RW"
 
+  val postcodeMap = listOf(
+    postcode(RELEASE_AREA_POSTCODE, 0.0, 0.0),
+    postcode(abcConstructionApprentice.postcode, 25000.0, 25000.0),
+    postcode(amazonForkliftOperator.postcode, 22800.0, 22800.0),
+    postcode(tescoWarehouseHandler.postcode, 1100.0, 1100.0),
+    postcode("LS110AN", 429017.0, 431869.0),
+    postcode("LS11 0AN", 429017.0, 431869.0),
+    postcode("M4 5BD", 385003.00, 398558.00),
+    postcode("NW1 6XE", 527870.40, 182081.17),
+    postcode("NG1 1AA", 457804.00, 340087.00),
+    postcode("E1 6AN", 533397.00, 181741.00),
+  ).associateBy({ it.code }, { it })
+
+  private fun postcode(postcode: String, xCoordinate: Double, yCoordinate: Double) =
+    Postcode(EntityId(), postcode, xCoordinate, yCoordinate)
+
   class Builder {
     var id: EntityId = EntityId(randomUUID().toString())
     var code: String = ""
